requirements.txt
# core
numpy
matplotlib
pillow
h5py
scikit-image
funcsigs
opencv-python
moviepy
torch
torchvision
tensorboard
tensorboardX
gym==0.15.4
pandas
ipdb
omegaconf
hydra-core
sklearn
more_tools

# for minecraft dataset
pyyaml

# RL
wandb
# mpi4py
transformers
pytorch-lightning

black
pylint

ai2thor